ia-service/requirements.txt
# =============== API ==================
fastapi==0.108.0
uvicorn[standard]==0.25.0
requests==2.32.5

# =============== Celery ===============
celery==5.3.4
redis==5.0.1
# =================Vector Store===========
faiss-cpu==1.7.4  # ← Ajouter cette ligne

# =============== Utils ================
python-dotenv==1.0.0

# =============== DB ===================
sqlalchemy==2.0.45
#mysqlclient==2.2.0  
pymysql==1.1.2

# =============== PDF ==================
pypdf==6.5.0

# =============== LangChain ============
langchain==1.2.0
langchain-community==0.4.1
langchain-groq==1.1.1
langchain-huggingface==1.2.0
cryptography

# =============== LLM / Embeddings =====
groq>=0.30.0,<1.0.0  # ← CHANGER ICI
sentence-transformers==5.2.0
torch==2.9.1
transformers==4.57.3
#=============Pic analys Package =========
pytesseract==0.3.13
pillow==12.0.0

# =============== ML deps ==============
# numpy==2.4.0
numpy<2.0.0,>=1.26.0
#scikit-learn>=1.4.0